Initial code and test checkin
The pycocontainer framework is a small "dependency injection" or
"inversion of control" framework for python.  It aims to help
decouple the administration of component lifecycles and dependency
relationships between components in a system from the components
themselves.  By doing so, it fosters more modular, flexible, and
decoupled object oriented software.